requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
pillow
opencv-python
numpy
pandas
scikit-learn
tqdm
python-multipart